made calcualtions classpath
implement the BillCalculator class to handle calculating the subtotal, tax, tip, and total for an Array of BillItem objects. Each calculation should be implemented as a separate static method which accepts an Array of BillItem Objects as input. You should assume that the Array will be filled left right (lowest index first), and that any unfilled space will be indicated with a null.